+++ pkg/ddl/BUILD.bazel
@@ -257,7 +257,6 @@ go_test(
         "//pkg/config",
         "//pkg/ddl/copr",
         "//pkg/ddl/ingest",
-        "//pkg/ddl/ingest/testutil",
         "//pkg/ddl/internal/session",
         "//pkg/ddl/logutil",
         "//pkg/ddl/placement",
