+++ pkg/ddl/ingest/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "mem_root.go",
         "message.go",
         "mock.go",
+        "util.go",
     ],
     importpath = "github.com/pingcap/tidb/pkg/ddl/ingest",
     visibility = ["//visibility:public"],
@@ -34,6 +35,7 @@ go_library(
         "//pkg/lightning/errormanager",
         "//pkg/lightning/log",
         "//pkg/meta",
+        "//pkg/metrics",
         "//pkg/parser/mysql",
         "//pkg/parser/terror",
         "//pkg/sessionctx",
