+++ tests/realtikvtest/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/config",
-        "//pkg/ddl/ingest/testutil",
+        "//pkg/ddl/ingest",
         "//pkg/domain",
         "//pkg/kv",
         "//pkg/session",
